pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "alter-ego-01"
version = "0.1.0"
description = "Personal AI Assistant Chatbot representing Aayush Vijayvergiya"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Aayush Vijayvergiya", email = "your-email@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "python-dotenv>=1.0.0",
    "openai>=1.108.1",
    "requests>=2.32.0",
    "pypdf>=6.1.0",
    "gradio>=5.46.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "html2text>=2024.2.26",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[project.scripts]
alter-ego = "src.alter_ego.app:create_app"

[tool.hatch.build.targets.wheel]
packages = ["src/alter_ego"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
